rename fuzz_constants to constants
